+++ src/entrypoints/recorder.ts
@@ -1,5 +1,4 @@
-// @ts-expect-error - Patched to include the version
-import { record as rrwebRecord, version } from 'rrweb'
+import { record as rrwebRecord } from '@rrweb/record'
 import { getRecordConsolePlugin } from '@rrweb/rrweb-plugin-console-record'
 
 // rrweb/network@1 code starts
@@ -670,14 +669,14 @@ export const getRecordNetworkPlugin: (options?: NetworkRecordOptions) => RecordP
 
 assignableWindow.__PosthogExtensions__ = assignableWindow.__PosthogExtensions__ || {}
 assignableWindow.__PosthogExtensions__.rrwebPlugins = { getRecordConsolePlugin, getRecordNetworkPlugin }
-assignableWindow.__PosthogExtensions__.rrweb = { record: rrwebRecord, version: 'v2', rrwebVersion: version }
+assignableWindow.__PosthogExtensions__.rrweb = { record: rrwebRecord, version: 'v2' }
 
 // we used to put all of these items directly on window, and now we put it on __PosthogExtensions__
 // but that means that old clients which lazily load this extension are looking in the wrong place
 // yuck,
 // so we also put them directly on the window
 // when 1.161.1 is the oldest version seen in production we can remove this
-assignableWindow.rrweb = { record: rrwebRecord, version: 'v2', rrwebVersion: version }
+assignableWindow.rrweb = { record: rrwebRecord, version: 'v2' }
 assignableWindow.rrwebConsoleRecord = { getRecordConsolePlugin }
 assignableWindow.getRecordNetworkPlugin = getRecordNetworkPlugin