+++ src/utils/globals.ts
@@ -52,7 +52,7 @@ interface PostHogExtensions {
         wrapOnError: (captureFn: (props: Properties) => void) => () => void
         wrapUnhandledRejection: (captureFn: (props: Properties) => void) => () => void
     }
-    rrweb?: { record: any; version: string; rrwebVersion: string }
+    rrweb?: { record: any; version: string }
     rrwebPlugins?: { getRecordConsolePlugin: any; getRecordNetworkPlugin?: any }
     canActivateRepeatedly?: (survey: any) => boolean
     generateSurveys?: (posthog: PostHog) => any | undefined
@@ -68,7 +68,7 @@ interface PostHogExtensions {
     }
     initDeadClicksAutocapture?: (
         ph: PostHog,
-        config?: DeadClicksAutoCaptureConfig
+        config: DeadClicksAutoCaptureConfig
     ) => LazyLoadedDeadClicksAutocaptureInterface
 }